requirements.txt
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-Classful==0.14.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==4.21.2
pycparser==2.21
PyJWT==2.4.0
regex==2022.6.2
requests==2.28.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0